requirement.txt
tensorflow
scikit-learn
flask
pillow
matplotlib
numpy
